+++ src/handlers/AbstractHandler.ts
@@ -211,10 +211,10 @@ export abstract class AbstractHandler implements ActionHandler {
       fileHistory.cleanup();
    }
 
-   private async setTargetFilePermissions(destPath: string, perms: PermissionModel, isDir = false) {
+   private async setTargetFilePermissions(destFile: string, perms: PermissionModel, isDir = false) {
       const mode = isDir ? perms.dirMode : perms.fileMode;
-      if (perms.uid >= 0 && perms.gid >= 0) await fsx.chown(destPath, perms.uid, perms.gid);
-      if (mode >= 0) await fsx.chmod(destPath, mode);
+      if (perms.uid >= 0 && perms.gid >= 0) await fsx.chown(destFile, perms.uid, perms.gid);
+      if (mode >= 0) await fsx.chmod(destFile, mode);
    }
 
    private getFolderPermissionPromises(ctx: RunnerContext, dirPath: string, folderPromises: PromiseMap): void {