+++ src/tasks/JobScheduler.ts
@@ -3,9 +3,10 @@ import { JobModel } from "../models/JobModel.js";
 import { JobRunnerSetup } from "../models/JobRunnerSetup.js";
 import { AbstractTask } from "./AbstractTask.js";
 import type { Job } from "../types/Config.types.js";
-import type { RunnerResult } from "../types/Task.types.js";
+import type { RunnerResult, TaskInfo } from "../types/Task.types.js";
 import type { RunnerOptions } from "../types/Options.types.js";
-import { JobError } from "../errors/JobError.js";
+
+type SchedulingInfo = Job & TaskInfo;
 
 export class JobScheduler extends AbstractTask<void> {
    protected runnerSetup: JobRunnerSetup;
@@ -35,6 +36,14 @@ export class JobScheduler extends AbstractTask<void> {
       }
    }
 
+   public pauseAll() {
+      for (const runner of this.runnerMap.values()) runner.resume();
+   }
+
+   public resumeAll() {
+      for (const runner of this.runnerMap.values()) runner.resume();
+   }
+
    public unscheduleAll() {
       for (const runner of this.runnerMap.values()) runner.unschedule();
       this.runnerMap = new Map<string, JobRunner>();
@@ -55,15 +64,11 @@ export class JobScheduler extends AbstractTask<void> {
    /**
     * Schedules a single job
     * @param job job definition
-    * @param defaults defaults to use if not defined in job definition
     * @throws JobError if job definition is not valid
     * @throws JobError if Job config is not valid
     */
-   public scheduleJob(job: Job, defaults = {}) {
+   public scheduleJob(job: Job) {
       let rescheduled = false;
-
-      // throw error if job is disabled
-      if (job.enabled === false) JobError.throw(job.id, `Cannot schedule disabled job [${job.id}]!`);
 
       // validate job --> throws JobError
       this.runnerSetup.validateJob(job);
@@ -75,11 +80,12 @@ export class JobScheduler extends AbstractTask<void> {
       }
 
       // create new job runner task
-      const task = new JobRunner(new JobModel(job, defaults), this.runnerSetup);
+      const task = new JobRunner(new JobModel(job), this.runnerSetup);
       this.runnerMap.set(job.id, task);
 
       // register events
       task.onScheduled(() => this.events.emit("job-scheduled", job, rescheduled));
+      task.onExecute(() => this.events.emit("job-execute", job));
       task.onStarted(() => this.events.emit("job-started", job));
       task.onFinished((stat: RunnerResult) => this.events.emit("job-finished", job, stat));
       task.onActivity((activity: string, path: string, count?: number) => this.events.emit("job-activity", job, activity, path, count));
@@ -87,6 +93,9 @@ export class JobScheduler extends AbstractTask<void> {
 
       // start cron scheduler for this job
       task.schedule();
+
+      // if job is disabled => pause job execution
+      if (job.enabled === false) task.pause();
 
       // mark as changed
       this.changed = true;
@@ -105,6 +114,10 @@ export class JobScheduler extends AbstractTask<void> {
       return this.runnerMap.has(jobId);
    }
 
+   public getJob(jobId: string): Job | undefined {
+      return this.runnerMap.get(jobId)?.job;
+   }
+
    public executeJob(jobId: string) {
       if (!this.runnerMap.has(jobId)) throw new Error(`Unknown job [${jobId}]!`);
       this.runnerMap.get(jobId)?.execute();
@@ -114,8 +127,32 @@ export class JobScheduler extends AbstractTask<void> {
       this.runnerSetup.validateJob(job);
    }
 
+   public pauseScheduling() {
+      for (const task of this.runnerMap.values()) this.pauseJobScheduling(task.job.id);
+      this.changed = true;
+   }
+
+   public resumeScheduling() {
+      for (const task of this.runnerMap.values()) this.resumeJobScheduling(task.job.id);
+      this.changed = true;
+   }
+
+   public pauseJobScheduling(jobId: string) {
+      const task = this.runnerMap.get(jobId);
+      if (task) task.pause();
+   }
+
+   public resumeJobScheduling(jobId: string) {
+      const task = this.runnerMap.get(jobId);
+      if (task && task.job.enabled !== false) task.resume();
+   }
+
    public getScheduledJobs(): Job[] {
       return Array.from(this.runnerMap.values()).map((runner) => runner.job);
+   }
+
+   public getScheduledJobsInfo(): SchedulingInfo[] {
+      return Array.from(this.runnerMap.values()).map((runner) => Object.assign({}, runner.job, runner.getInfo()));
    }
 
    public override async gracefulTerminate(timeout: number = 1000) {
@@ -140,6 +177,15 @@ export class JobScheduler extends AbstractTask<void> {
     */
    public onJobScheduled(cb: (job: Job, rescheduled: boolean) => void) {
       this.events.on("job-scheduled", cb);
+   }
+
+   /**
+    * Emitted if manual job execution has been started via executeJob().
+    * Note: This event is not emitted, if the job is paused or already running
+    * @param cb callback
+    */
+   public onJobExecute(cb: (job: Job) => void) {
+      this.events.on("job-execute", cb);
    }
 
    /**